pyproject.toml
[project]
name = "fileskadis"
version = "0.1.0"
description = "A file operations library for PDF/Image aggregation, separation, and masking"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "fileskadis", email = "fileskadis@example.com" }]
keywords = ["pdf", "image", "aggregation", "masking", "redaction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pypdfium2>=4.30.0",
    "pillow>=10.0.0",
    "structlog>=24.0.0",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
fileskadis-ui = "fileskadis.ui.app:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.resolution]
resolver = "uv"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

